refactor: Remove TelemetryExceptionStatusService and ExceptionService
