+++ Common/Server/Infrastructure/Postgres/SchemaMigrations/Index.ts
@@ -87,6 +87,5 @@ export default [
   MigrationName1723825511054,
   MigrationName1723828588502,
   MigrationName1724078044172,
-  MigrationName1720812937067,
   MigrationName1724610006927,
 ];
