+++ Common/Server/Services/Index.ts
@@ -130,7 +130,8 @@ import AnalyticsBaseModel from "Common/Models/AnalyticsModels/AnalyticsBaseModel
 import CopilotPullRequestService from "./CopilotPullRequestService";
 import ServiceCatalogDependencyService from "./ServiceCatalogDependencyService";
 import TelemetryAttributeService from "./TelemetryAttributeService";
-import TelemetryExceptionStatusService from "./TelemetryExceptionStatusService";
+import TelemetryExceptionService from "./TelemetryExceptionService";
+import ExceptionInstanceService from "./ExceptionInstanceService";
 
 const services: Array<BaseService> = [
   AcmeCertificateService,
@@ -273,7 +274,7 @@ const services: Array<BaseService> = [
   ServiceCopilotCodeRepositoryService,
   CopilotPullRequestService,
 
-  TelemetryExceptionStatusService,
+  TelemetryExceptionService,
 ];
 
 export const AnalyticsServices: Array<
@@ -284,6 +285,7 @@ export const AnalyticsServices: Array<
   MetricService,
   MonitorMetricsByMinuteService,
   TelemetryAttributeService,
+  ExceptionInstanceService
 ];
 
 export default services;
